Switch to authlib for user password reset token generation and verification
In version 2.0.0 (released 2021-05-11) itsdangerous deprecated support for TimedJSONWebSignatureSerializer and in version 2.1.0 it was removed from the codebase. https://itsdangerous.palletsprojects.com/en/2.1.x/changes/?highlight=timedjsonwebsignatureserializer#version-2-1-0

This commit switches to authlib as was suggested by https://github.com/pallets/itsdangerous/issues/129

Looking through the commits in keg-auth that did the same was useful for research, but did not map 1:1 with the existing implementation. https://github.com/level12/keg-auth/commit/e96ac2e5b592d31dec29cfb772a5e9c34314be32

As part of this change, the user_id must be passed so that an identical salt will be generated in order to validate the password reset token.

In order to assist users in staying up-to-date with dependencies, a check during application startup for authlib has been added.

Also, the CHANGELOG was updated for this change.