:sparkles: Split Wizard and App to facilitate alternative deployments and application lifecycles (such as Docker)
This includes removing the outboard version.txt, and instead relying on the alembic version and running upgrades in the wizard every time it is executed.

A tab was added to the settings page to display these two different version numbers.

Also, the mechanisms of the wizard was split out into function so that it may be programmatically driven in the future.